requirements.txt
streamlit
python-docx
PyPDF2
scikit-learn
